update: ajustes na api

+++ src/main/java/com/pol/avaliacao/services/ProcessoService.java
@@ -1,57 +1,73 @@
 package com.pol.avaliacao.services;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.stereotype.Service;
 
 import com.pol.avaliacao.entities.Processo;
 import com.pol.avaliacao.entities.ProcessoDTO;
 import com.pol.avaliacao.repositories.ProcessoRepository;
 
-import jakarta.websocket.server.ServerEndpoint;
-
 @Service
 public class ProcessoService {
 	@Autowired
 	private ProcessoRepository repository;
 
-	public ProcessoDTO createProcesso(ProcessoDTO processoDTO) {
+	public ProcessoDTO createProcesso(ProcessoDTO processoDto) {
+		if (repository.existsById(processoDto.getNumero_processo())) {
+			throw new IllegalArgumentException("O processo com este número já existe.");
+		}
+
 		Processo processo = new Processo();
-		processo.setNumero_processo(processoDTO.getNumero_processo());
-		processo.setAutor(processoDTO.getAutor());
-		processo.setReu(processoDTO.getReu());
-		processo.setStatus(processoDTO.getStatus());
-		processo.setData_criacao(processoDTO.getData_criacao());
-		return processoDTO;
+		processo.setNumero_processo(processoDto.getNumero_processo());
+		processo.setAutor(processoDto.getAutor());
+		processo.setReu(processoDto.getReu());
+		processo.setStatus(processoDto.getStatus());
+		processo.setData_criacao(processoDto.getData_criacao());
+		repository.save(processo);
+		return processoDto;
 	}
-	
-	public ProcessoDTO updateProcesso(ProcessoDTO processoDTO, Integer processoId) {
+
+	public ProcessoDTO updateProcesso(ProcessoDTO processoDto, Integer processoId) {
 		Processo processoDb = repository.getReferenceById(processoId);
-		processoDb.setReu(processoDTO.getReu());
-		processoDb.setAutor(processoDTO.getAutor());
-		processoDb.setStatus(processoDTO.getStatus());
-		return processoDTO;
+		processoDb.setNumero_processo(processoDto.getNumero_processo());
+		processoDb.setReu(processoDto.getReu());
+		processoDb.setAutor(processoDto.getAutor());
+		processoDb.setStatus(processoDto.getStatus());
+
+		repository.save(processoDb);
+		return processoDto;
 	}
-	
-	  private ProcessoDTO converter (Processo processo) {
-	        ProcessoDTO result = new ProcessoDTO();
-	        result.setNumero_processo(processo.getNumero_processo());
-	        result.setReu(processo.getReu());
-	        result.setAutor(processo.getAutor());
-	        result.setStatus(processo.getStatus());
-	        result.setData_criacao(processo.getData_criacao());
-	        return result;
-	    }
-	
-	
-	public List<ProcessoDTO> getAll(){
+
+	private ProcessoDTO converter(Processo processo) {
+		ProcessoDTO result = new ProcessoDTO();
+		result.setNumero_processo(processo.getNumero_processo());
+		result.setReu(processo.getReu());
+		result.setAutor(processo.getAutor());
+		result.setStatus(processo.getStatus());
+		result.setData_criacao(processo.getData_criacao());
+		return result;
+	}
+
+	public List<ProcessoDTO> getAll() {
 		return repository.findAll().stream().map(this::converter).collect(Collectors.toList());
 	}
-	
+
+	public Optional<Processo> getProcesso(ProcessoDTO processoDto, Integer processoId) {
+		return repository.findById(processoId);
+
+	}
+
 	public String deleteProcesso(Integer processoId) {
+		if (!repository.existsById(processoId)) {
+			throw new IllegalArgumentException("O processo com este número não existe no banco de dados.");
+		}
 		repository.deleteById(processoId);
 		return "Processo deletado!";
 	}
+
 }
